Using ldd to track external libraries required by a binary package.
This patch adds ldd support to the binary packages:

A file ending with '.extlib' is created along with each binary package, listing
the external libraries required by the binaries contained in this package, as
given by a ldd call on each of these binaries.

When trying to install a binary package, the required external libraries listed
in the '.extlib' file are fed to the 'whereis' utility in order to check their
presence on the system. The binary package is installed only if the check
succeeds.